do not set version in config
